Added 'ProcessMultiKey' option, enabled by default
If enabled, when the entire section content is requested, it'll be returned with multiple keys present. Otherwise only a single instance of each key will be read.
The option has no bearing for single value operations, for them only the first value associated with the requested key will be returned. Write operations will replace all keys with the matching name.

In theory this change can fix some inconsistencies with 'GetPrivateProfileSection' functions.